dmsetup: add support to output devices in bootformat style
Add a new parameter to be able to output the bootformat of one or more
devices. The bootformat for a device can be obtained by doing a:

 dmsetup table <device...> --bootformat

where device can be a specific device, a list of the devices or empty
for all devices. The output will look like this:

 <dev_name>,<uuid>,<flags>,<table>[,<table>+][;<dev_name>,<uuid>,<flags>,<table>[,<table>+]]+

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>