libdm: Add dm_find_unescaped_char to libdevmapper.
Add helper functions to split a string into tokens ignoring escaped
chars. The function finds the first token in the string that is delimited
by one non escaped char. In case it founds an escaped token finds next
token.

When a token is found the string is terminated by overwriting the
delimiter with a null byte ('\0') and the pointer to the search string
is updated to point past the token ready for the next call.

In case no delimiter was found, the token is taken to be the entire
string.

Aditionally there is dm_unescape_colons and dm_unescape_semicolons
helpers to unescape these characters in situ, it replaces all occurrences
of "\," or "\;" with ',' or ';'. This is normally used to unescape colons
and semi-colons used in boot format.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>